Changed behavior for file extensions:
If there is a standart extension given by the user (meaning 'someFileName' followed by a '.' and some letters), the file is created with the extension typed in by the user.

However: If the user does something like 'hello. txt3 e', the file will be automatically created as 'hello. txt3 e.txt'.

I think this makes more sense.
The primary "customer" of the app is a developer, who knows what they are doing, and can name their files accordingly.
If someone else uses the app, they don't have to use file extensions, and they can't "mess up" ;-)